requirements.txt
langchain==0.1.13
langchain-openai==0.1.5
langchain-community==0.0.28
faiss-cpu==1.7.4
pandas==2.0.3
python-dotenv==1.0.0
openai==1.3.9
